+++ sympl/__init__.py
@@ -21,7 +21,8 @@ from ._core.util import (
     get_component_aliases)
 from ._core.state import (
     get_numpy_arrays_with_properties,
-    restore_data_arrays_with_properties)
+    restore_data_arrays_with_properties,
+    initialize_numpy_arrays_with_properties)
 from ._components import (
     PlotFunctionMonitor, NetCDFMonitor, RestartMonitor,
     ConstantPrognostic, ConstantDiagnostic, RelaxationPrognostic,
@@ -44,6 +45,7 @@ __all__ = (
     get_numpy_array, jit,
     restore_dimensions, get_numpy_arrays_with_properties,
     restore_data_arrays_with_properties,
+    initialize_numpy_arrays_with_properties,
     get_component_aliases,
     PlotFunctionMonitor, NetCDFMonitor, RestartMonitor,
     ConstantPrognostic, ConstantDiagnostic, RelaxationPrognostic,